+++ Smithtry1000/Line.cpp
@@ -1,17 +1,17 @@
 ﻿#include "Line.h"
 
-void LinesElement::SetAlpha(double alpha)
+void LinesElement::SetAlpha(long double alpha)
 {
 	this->_alpha = alpha;
 }
 
-double LinesElement::GetAlpha()
+long double LinesElement::GetAlpha()
 {
 	return this->_alpha;
 }
 
-LinesElement::LinesElement(mode mode, double value, double frequency, Point point, map<chartMode, tuple<double, double>> chartParameters,
-	map<parameterMode, Complex> parameters, double electricalLength, double mechanicalLength, double theta, double lambda, double alpha) :
+LinesElement::LinesElement(mode mode, long double value, long double frequency, Point point, map<chartMode, tuple<long double, long double>> chartParameters,
+	map<parameterMode, Complex> parameters, long double electricalLength, long double mechanicalLength, long double theta, long double lambda, long double alpha) :
 	VerticalLinesElement(mode, value, frequency, point, chartParameters, parameters, electricalLength, mechanicalLength, theta, lambda)
 {
 	this->SetAlpha(alpha);
